api: move socket api views to api/socket.py
