+++ app.py
@@ -2,7 +2,6 @@ from flask import Flask
 from flask import jsonify
 from flask import render_template
 from flask import request
-from flask import session
 from pymongo import MongoClient
 import flask_socketio as io
 
@@ -12,39 +11,6 @@ app = Flask(__name__)
 settings.configure(app)
 socketio = io.SocketIO(app)
 db_client = MongoClient(app.config['DATABASE_URL'])[app.config['DATABASE_NAME']]
-
-
-def chat(message, room):
-    io.emit('chat', message, room=room)
-
-
-@socketio.on('chat')
-def on_chat(message):
-    print('received message: ' + str(message))
-    if 'room' in session:
-        chat(message, session['room'])
-    else:
-        # chat to a global chat room?
-        print('No room stored on session')
-
-
-@socketio.on('join')
-def on_join(data):
-    username = data['username']
-    room = data['room']
-    print('join {}'.format((username, room)))
-    io.join_room(room)
-    session['room'] = room
-    io.emit(username + ' has entered the room.', room=room)
-
-
-@socketio.on('leave')
-def on_leave(data):
-    username = data['username']
-    room = data['room']
-    print('leave {}'.format((username, room)))
-    io.leave_room(room)
-    io.emit(username + ' has left the room.', room=room)
 
 
 @app.route('/')
